Brody/docs versions (#125)
## Summary

Preview link:
https://langchain-5e9cc07a-preview-brodyd-1754925111-e44a21e.mintlify.app/oss/python/overview

This PR does two main things:

1. It reworks the `builder.py` file to generate JS and Python versions
of the docs.
2. It adds CSS and JS files for conditionally hiding the language
picker.

The requirements for the picker are:
- Show the picker when viewing the LangGraph docs
- Hide the picker when viewing the LangGraph Platform or LangChain Labs
docs

## Review requests + info

We're doing custom things on top of Mintlify's stuff. It would be great
if someone can help verify the changes to the `builder.py` file are
acceptable, and although it's not the most elegant solution, that the
CSS and JS files are an acceptable though likely temporary solution.

I tried to update and fix tests appropriately but I really need someone
to validate those changes are good. I haven't fixed/updated unit tests
like this before.

## Todos

- [x] Get this PR reviewed and implement any feedback
- [x] Fix tests
- [ ] Merge with https://github.com/langchain-ai/docs/pull/109
- [ ] 🎉

---------

Co-authored-by: Lauren Hirata Singh <lauren@langchain.dev>