Update agentic-rag.mdx: rewrite_question Node: Role and Content based dictionary replaced with LangChain Human Message
As per the docs live officially, in the rewrite_question node, the response of the model for rewriting the query is returned in dictionary format to messages in the state.

return {"messages": [{"role": "user", "content": response.content}]}

I understand this has been done keeping in mind to modify the message's role from ai/assistant to user in the state. While this is not critical for the time being, it might cause errors while printing graph's streaming updates.

for node, update in chunk.items():
        print("Update from node", node)
        update["messages"][-1].pretty_print()
        print("\n\n")

This might have been skipped due to running an example where rewrite node was not required.

However, if the rewrite node comes into play, it would push the dictionary-based format of messages and not the standard Messages class format. The pretty print function is valid only on Message classes and not a dictionary; this would lead to an attribute error: 
AttributeError: 'dict' object has no attribute 'pretty_print'

For AI learners, especially those who have been exploring LangChain v1.0 docs might get blocked due to this while following the tutorial. 

Please merge this one so that it adds to the quality and trust for LangChain docs.